Assignment5A completed
The level has some parts where you might "fall through the floor." This is intentional and part of the level design! There are some parts where you might get stuck on the slopes. This is not intentional but will fix itself if you take your hands off the keyboard for a moment.